+++ next.config.mjs
@@ -21,7 +21,19 @@ const nextConfig = {
         ]
     },
     reactStrictMode: false,
-    trailingSlash: false
+    trailingSlash: false,
+    eslint: {
+        // Warning: This allows production builds to successfully complete even if
+        // your project has ESLint errors.
+        ignoreDuringBuilds: true,
+      },
+    typescript: {
+    // !! WARN !!
+    // Dangerously allow production builds to successfully complete even if
+    // your project has type errors.
+    // !! WARN !!
+    ignoreBuildErrors: true,
+    },
 };
 
 export default nextConfig;